Alterações feitas nas classes para a função pedido

+++ src/main/java/br/com/robotrading/web/dao/RobosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.robotrading.web.model.Robo;
 
-public interface RobosDAO extends JpaRepository<Robo, Long> {
+public interface RobosDAO extends JpaRepository<Robo, Integer> {
 	
 }
